Fix automatic torch lighting checks to sample only air
Exclude occupied cells from light queries and the averaging divisor so surrounding stone cannot make lit air appear dark. Preserve fractional light levels, skip areas without air samples, and report the sampled light when placing torches.

Validation: 23 lighting tests passed; the isolated commit sources also compile and pass all 8 baseline and regression tests.

+++ src/client/java/ai/moeru/airicraft/agent/lighting/LightingPolicyEvaluator.java
@@ -15,8 +15,8 @@ public final class LightingPolicyEvaluator {
 		boolean supportedActivity,
 		boolean torchAvailable,
 		boolean skyVisible,
-		int combinedLightLevel,
-		int blockLightLevel,
+		double combinedLightLevel,
+		double blockLightLevel,
 		boolean nearbyTorch
 	) {
 		if (policy == null || !policy.enabled() || !supportedActivity || !torchAvailable || nearbyTorch) {
@@ -25,7 +25,7 @@ public final class LightingPolicyEvaluator {
 		if (policy.requireUnderground() && skyVisible) {
 			return false;
 		}
-		int observedLight = policy.mode() == LightingPolicy.Mode.SPAWN_PROOF
+		double observedLight = policy.mode() == LightingPolicy.Mode.SPAWN_PROOF
 			? blockLightLevel
 			: combinedLightLevel;
 		return observedLight <= policy.maxLightLevel();
